change structure : add enumeration and refactor folders (#16)
