+++ src/main/java/com/alkemy/wallet/enumeration/Currency.java
@@ -1,0 +1,9 @@
+package com.alkemy.wallet.enumeration;
+
+public class Currency {
+  public enum currency {
+    ARS,
+    USD
+  }
+
+}
